platform/deployment/paddleocr/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
paddlepaddle==2.6.1
paddleocr==2.7.3
python-multipart==0.0.9
Pillow==10.3.0
numpy==1.26.4
